% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_colourise_text.R
\name{cat_error}
\alias{cat_error}
\title{Print "ERROR" in Light Red Color}
\usage{
cat_error(msg = "ERROR\\n", bg = NULL)
}
\arguments{
\item{msg}{The text message to be printed. Defaults to "ERROR".}

\item{bg}{The background color. Defaults to NULL.}
}
\description{
This function prints "ERROR" in light red color.
}
