% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_colourise_text.R
\name{cat_headline}
\alias{cat_headline}
\title{Print Text in White Color (Headline Style)}
\usage{
cat_headline(msg, bg = NULL)
}
\arguments{
\item{msg}{The text message to be printed.}

\item{bg}{The background color. Defaults to NULL.}
}
\description{
This function prints text in white color (headline style).
}
