% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_misc.R
\name{check_error}
\alias{check_error}
\title{Check for Errors}
\usage{
check_error(
  err,
  expr_ok = {
     cat_ok()
 },
  expr_err = {
     cat_error()
 }
)
}
\arguments{
\item{err}{Any Type. In case of an error occurred it must contain try-error as class}

\item{expr_ok}{An expression. This runs in case of no error.}

\item{expr_err}{An expression. This runs in case of an error.}
}
\value{
err
}
\description{
Check for Errors
}
