% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{combineDataTables}
\alias{combineDataTables}
\title{Combine a List of data.tables into a single large data.table}
\usage{
combineDataTables(dt_list)
}
\arguments{
\item{dt_list}{A list of data.tables to be combined.}
}
\value{
A single data.table containing the combined data from the input list.
}
\description{
This function takes a list of data.tables with the same column names
and combines them into a single large data.table. It ensures that only
common columns are included in the resulting data.table.
}
\examples{
# Create a list of data.tables with common columns
library(data.table)
dt1 <- data.table(ID = 1:3, Value = c(10, 20, 30))
dt2 <- data.table(ID = 4:6, Value = c(40, 50, 60))
dt_list <- list(dt1, dt2)

# Combine the data.tables into a single data.table
big_dt <- combineDataTables(dt_list)
print(dt1)
print(dt2)
print(big_dt)

}
