% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_dir.R
\name{create_dirs}
\alias{create_dirs}
\title{This function is used by the framework itself}
\usage{
create_dirs(project_name = PROJECT_NAME)
}
\arguments{
\item{project_name}{name of the project}
}
\description{
This function is used by the framework itself
}
