% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_db.R
\name{deleteTableContentFromDatabase}
\alias{deleteTableContentFromDatabase}
\title{Delete all rows from a table in the database.}
\usage{
deleteTableContentFromDatabase(db_connection, table_name)
}
\arguments{
\item{db_connection}{A database connection object.}

\item{table_name}{The name of the table from which rows should be deleted.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{         The result of the delete operation.
}\if{html}{\out{</div>}}
}
\description{
This function deletes all rows from a specified table in the database.
}
\examples{
\dontrun{
# Connect to a database (replace with your actual connection details)
con <- DBI::dbConnect(RSQLite::SQLite(), dbname = "your_database.db")

# Delete all rows from the "your_table" table
deleteTableContentFromDatabase(con, "your_table")
}

}
