% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_helper_functions.R
\name{filterResources}
\alias{filterResources}
\title{Filters the given resources table by the given filter patterns.}
\usage{
filterResources(resources, filter_patterns)
}
\arguments{
\item{resources}{A data.table representing the resources table to be filtered.}

\item{filter_patterns}{A list of filter conditions. Each condition is a character string containing multiple
subconditions separated by '+'.}

\item{row}{A row (list or data.frame) to be checked against the condition.}

\item{condition}{A list where each element is a grep pattern, and the name corresponds to the column in the row.}
}
\value{
A filtered data.table based on the given filter patterns.

TRUE if the row fulfills the condition, FALSE otherwise.
}
\description{
This function filters a resources table based on the provided filter patterns.
}
\details{
This function applies an OR operation across the filter patterns, meaning that a row will be retained if at
least one condition is fulfilled.
However, within each individual condition (subconditions separated by '+'), an AND operation is applied,
requiring all subconditions to be met for the condition to be satisfied.
}
