% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_helper_functions.R
\name{flatten_list}
\alias{flatten_list}
\title{Flatten a nested list}
\usage{
flatten_list(x, prefix = NULL)
}
\arguments{
\item{x}{The input nested list.}

\item{prefix}{An optional prefix to be added to the names of the flattened list.}
}
\value{
A flattened list.
}
\description{
This function takes a nested list and flattens it into a single-level list and
removes parent names from the variable names.
It is particularly useful when you have flattened nested structures and want to
clean up the variable names by removing the prefixes introduced during flattening.
}
\details{
The function recursively iterates through the nested list, creating a flattened list
with modified variable names by adding the specified prefix. If no prefix is provided,
the original names are retained. The function uses a regular expression to remove parent names from variable names.
It looks for patterns at the beginning of each name and removes everything up to
the first uppercase letter, assuming that uppercase letters indicate the start of a new variable.
For example, "retrieve_fhir_server_SORT" becomes "SORT".
}
