% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_misc.R
\name{get_pattern_YYYY_MM}
\alias{get_pattern_YYYY_MM}
\title{Get a regular expression pattern for matching YYYY-MM format}
\usage{
get_pattern_YYYY_MM()
}
\value{
A character vector representing the regular expression pattern for YYYY-MM format.
}
\description{
This function returns a regular expression pattern for matching the YYYY-MM format
(four-digit year followed by a hyphen and two-digit month). It can be used for
validating and extracting year and month information from strings.
}
\examples{
get_pattern_YYYY_MM()

}
