% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_helper_functions.R
\name{getVarByNameOrDefaultIfMissing}
\alias{getVarByNameOrDefaultIfMissing}
\title{Get the value of a variable by name or a default value if the variable is missing.}
\usage{
getVarByNameOrDefaultIfMissing(var_name, default = NA)
}
\arguments{
\item{var_name}{A character string specifying the name of the variable.}

\item{default}{The default value to be returned if the variable is missing. Defaults to NA.}
}
\value{
The value of the variable if it exists; otherwise, the specified default value.
}
\description{
This function checks if a variable with the specified name exists. If it does,
it returns the value of the variable; otherwise, it returns the specified default value.
}
\examples{
# Set a variable
my_variable <- 123

# Use getVarByNameOrDefaultIfMissing to retrieve the value
result <- getVarByNameOrDefaultIfMissing("my_variable", default = 42)
print(result)  # Output: 123

# Try with a non-existing variable
result_missing <- getVarByNameOrDefaultIfMissing("nonexistent_variable", default = "Not found")
print(result_missing)  # Output: "Not found"

}
