% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_helper_functions.R
\name{isError}
\alias{isError}
\title{Tests the passed object for being an error.}
\usage{
isError(obj)
}
\arguments{
\item{obj}{The object to be tested for being an error.}
}
\value{
TRUE if the object is an 'try-error', FALSE otherwise.
}
\description{
This function checks whether the input object is an 'try-error'. If the object
is an error, the function returns TRUE; otherwise, it returns FALSE.
}
\examples{
# Example 1: Check if an error is an error
result <- try(log("a"))
isError(result)

# Example 2: Check if a numeric value is not an error
numeric_result <- 42
isError(numeric_result)

}
