% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{isSimpleNA}
\alias{isSimpleNA}
\title{Check if a value is a simple NA.}
\usage{
isSimpleNA(x)
}
\arguments{
\item{x}{The value to be checked.}
}
\value{
TRUE if the value is a simple NA, otherwise FALSE.
}
\description{
This function checks whether a value is a simple NA by ensuring that it is atomic (simple),
has a length of 1, and the value is NA.
}
\examples{
x <- NA
isSimpleNA(x) # TRUE

y <- c(1, NA, 3)
isSimpleNA(y) # FALSE

}
