% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{isSimpleTrueOrNot0}
\alias{isSimpleTrueOrNot0}
\title{Check if a value is a simple true value (TRUE) or not zero (0).}
\usage{
isSimpleTrueOrNot0(x)
}
\arguments{
\item{x}{The value to be checked.}
}
\value{
TRUE if the input is a simple true value or a non-zero numeric value, FALSE otherwise.
}
\description{
This function evaluates whether the input value is an atomic element, has a length of 1,
and is either a logical true value (TRUE) or a non-zero numeric value. If these conditions are met,
the function returns TRUE; otherwise, it returns FALSE.
}
\examples{
# Example 1: Check if TRUE is a simple true value (TRUE)
isSimpleTrueOrNot0(TRUE)

# Example 2: Check if 0 is not a simple true value (FALSE)
isSimpleTrueOrNot0(0)

# Example 3: Check if 1 is a simple true value (TRUE)
isSimpleTrueOrNot0(1)

# Example 4: Check if a boolean vector is a simple true value (FALSE)
isSimpleTrueOrNot0(c(TRUE, TRUE))

}
