% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{isValidTable}
\alias{isValidTable}
\title{Check if a table is valid.}
\usage{
isValidTable(table)
}
\arguments{
\item{table}{The table to be checked.}
}
\value{
TRUE if the table is valid (not NULL, not NA, and has at least one row),
otherwise FALSE.
}
\description{
This function checks whether a table is valid by verifying that it is not NULL,
not NA, and has at least one row.
}
\examples{
# Check if a data.table is valid
library(data.table)
dt <- data.table(A = 1:3, B = c('X', 'Y', 'Z'))
isValid <- isValidTable(dt)
print(isValid)  # Output: TRUE

}
