% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_helper_functions.R
\name{makeRelative}
\alias{makeRelative}
\title{Extracts the local, relative part from absolute references.}
\usage{
makeRelative(references)
}
\arguments{
\item{references}{A character vector containing absolute references.}
}
\value{
A character vector with only the local, relative part of each reference.
}
\description{
This function takes a character vector of absolute references and removes the leading part
to return only the local, relative portion of each reference.
}
\examples{
absolute <- c('Patient/PID_001', 'Encounter/EID_001', 'Condition/CID_001')
relative <- makeRelative(absolute)
print(relative)

}
