% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_helper_functions.R
\name{namedListByParam}
\alias{namedListByParam}
\title{Creates a named list of all parameters. The names are the variable names of the parameters.}
\usage{
namedListByParam(...)
}
\arguments{
\item{...}{A vector from which a named list will be created.}
}
\description{
Creates a named list of all parameters. The names are the variable names of the parameters.
}
\examples{
# Create a named list with specified names using the variables
a <- 1
b <- 2
c <- 3
named_list <- namedListByParam(a, b, c)
print(named_list)  # Output: $a [1] 1, $b [1] 2, $c [1] 3

#Create a named list of 2 tables
library(data.table)

table1 <- data.table(
  column1 = c(1:5),
  column2 = c('A', 'B', 'C', 'D', 'E')
)
table2 <- data.table(
  column1 = c(1:10),
  column2 = c('f', 'g', 'h', 'i', 'j', 'k', 'l', 'm', 'n', 'o')
)

tableList <- namedListByParam(table1, table2)

}
