% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_misc.R
\name{polar_add_common_request_params}
\alias{polar_add_common_request_params}
\title{Add Common Parameters to FHIR Resource Request}
\usage{
polar_add_common_request_params(parameters = NULL)
}
\arguments{
\item{parameters}{A list of FHIR resource query parameters.}
}
\value{
A modified list of FHIR resource query parameters with common parameters added.
}
\description{
This function adds common parameters, such as '_count' and '_sort', to a list of FHIR resource query parameters.
}
