% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_dir.R
\name{polar_add_to_bundles_path}
\alias{polar_add_to_bundles_path}
\title{Return the combined paths of the bundles directory of the specific sub project and a new path}
\usage{
polar_add_to_bundles_path(path)
}
\arguments{
\item{path}{the sub directory or file name to add}
}
\value{
A character of length one containing the path to add to the sub project specific bundles directory.
}
\description{
Return the combined paths of the bundles directory of the specific sub project and a new path
}
