% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_misc.R
\name{polar_as_date}
\alias{polar_as_date}
\title{Convert a date representation to Date format}
\usage{
polar_as_date(date_column)
}
\arguments{
\item{date_column}{A column containing date information in a polar format.}
}
\value{
A Date vector representing the converted date information.
}
\description{
This function takes a date column in a polar format, cleans and standardizes
the format, and converts it to Date format. The input date_column is expected
to be in a format containing year, month, and day information. The function
supports patterns for YYYY and YYYY-MM. It utilizes the functions
\code{\link{get_pattern_YYYY}} and \code{\link{get_pattern_YYYY_MM}} for obtaining
regular expression patterns.
}
\examples{
library(lubridate)

# Test case 1: YYYY format
date_column_YYYY <- c("2022", "1990", "1980")
result_YYYY <- polar_as_date(date_column_YYYY)

# Test case 2: YYYY-MM format
date_column_YYYY_MM <- c("2022-12", "1990-05", "1980-11")
result_YYYY_MM <- polar_as_date(date_column_YYYY_MM)

# Test case 3: Date with time, should be cleaned
date_column_with_time <- c("2022-12-01T15:30:00", "1990-05-01T08:45:00")
result_with_time <- polar_as_date(date_column_with_time)

# Test case 4: Date with '/' separator, should be replaced with '-'
date_column_slash_separator <- c("2022/12/01", "1990/05/01")
result_slash_separator <- polar_as_date(date_column_slash_separator)

}
\seealso{
\code{\link{get_pattern_YYYY}}, \code{\link{get_pattern_YYYY_MM}}
}
