% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_sample_resources.R
\name{polar_download_and_crack_parallel}
\alias{polar_download_and_crack_parallel}
\title{Download and Crack FHIR Resources in Parallel}
\usage{
polar_download_and_crack_parallel(
  request = REQUEST_ENCOUNTER,
  table_description = TABLE_DESCRIPTION$Encounter,
  bundles_at_once = 20,
  verbose = 1,
  bundles_left = MAX_BUNDLES,
  max_cores = 2,
  log_errors = "enc_error.xml"
)
}
\arguments{
\item{request}{The initial FHIR request to start the download process.}

\item{table_description}{The description of the expected table structure for the resource.}

\item{bundles_at_once}{The maximum number of bundles to process in a single iteration.}

\item{verbose}{An integer specifying the verbosity level.}

\item{bundles_left}{The total number of bundles to download and process.}

\item{max_cores}{The maximum number of CPU cores to use for parallel processing.}

\item{log_errors}{The file name for logging errors during the process.}
}
\value{
A data.table containing the cracked FHIR resource data.
}
\description{
This function downloads FHIR resources by making parallel requests and cracks the bundles in parallel.
}
