% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_fhir_wrapper.R
\name{polar_fhir_search}
\alias{polar_fhir_search}
\title{Wrapper for fhircrackr::fhir_search with automatic Authentication}
\usage{
polar_fhir_search(
  request = fhircrackr::fhir_current_request(),
  body = NULL,
  max_bundles = MAX_BUNDLES,
  verbose = 1,
  max_attempts = 5,
  delay_between_attempts = 10,
  log_errors = NULL,
  save_to_disc = NULL,
  delay_between_bundles = 0
)
}
\arguments{
\item{request}{A character of length one containing a fhir search request.}

\item{body}{A  character of length one containing the body for a POST request.}

\item{max_bundles}{An integer of length one. How many bundles should be retrieved.}

\item{verbose}{An integer of length one. Selects the verbosity. Defaults to 1.}

\item{max_attempts}{An integer of length one. The number of attempts, if some error occurs.}

\item{delay_between_attempts}{An number of length one. The time between two attempts in seconds.}

\item{log_errors}{Either NULL or a character vector of length one indicating the name of a file
in which to save the http errors.
NULL means no error logging. When a file name is provided, the errors are saved in the specified file.
Defaults to NULL. Regardless of the value of log_errors the most recent http error message whithin
the current R session is saved internally and can be accessed with fhir_recent_http_error().}

\item{save_to_disc}{Either NULL or a character vector of length one indicating the name of a directory
in which to save the bundles. If a directory name is provided, the bundles are saved as numerated xml-files
into the directory specified and not returned as a bundle list in the R session.
This is useful when a lot of bundles are to be downloaded and keeping them all in one R session might
overburden working memory. When the download is complete, the bundles can be loaded into R using fhir_load().
Defaults to NULL, i.e. bundles are returned as a list within the R session.}

\item{delay_between_bundles}{A numeric scalar specifying a time in seconds to wait between pages
of the search result, i.e. between downloading the current bundle and the next bundle.
This can be used to avoid choking a weak server with too many requests to quickly. Defaults to zero.}
}
\value{
A fhir_bundle_list when save_to_disc = NULL (the default), else NULL.
}
\description{
Wrapper for fhircrackr::fhir_search with automatic Authentication
}
