% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_sample_resources.R
\name{polar_refresh_token}
\alias{polar_refresh_token}
\title{Refresh FHIR-Server authentication Token}
\usage{
polar_refresh_token()
}
\value{
A character string representing the refreshed FHIR-Server authentication token.
}
\description{
This function refreshes the FHIR-Server authentication token by making a request to the specified token refresh URL.
}
