% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_sample_resources.R
\name{polar_sample_resources}
\alias{polar_sample_resources}
\title{Sample Resources of certain IDs}
\usage{
polar_sample_resources(
  endpoint,
  resource,
  parameters = parameters,
  sample_size = 20,
  seed = as.double(Sys.time()),
  verbose = 1
)
}
\arguments{
\item{endpoint}{A character of length 1 containing the FHIR R4 endpoint.}

\item{resource}{A character of length 1 containing the Resource's name.}

\item{parameters}{Either a character of length 1 containing the parameter string or
a named vector of the form c("gender" = "male", "_summary" = "count").}

\item{sample_size}{A integer of length 1 containing the size of the sample.}

\item{seed}{A integer of length 1 containing the seed for the random generator.}

\item{verbose}{An integer of length 1 containing the level of verbosity. Defaults to 1.}
}
\value{
A list of bundles containing sampled resources.
}
\description{
Sample Resources of certain IDs
}
