% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_dir.R
\name{polar_save_rdata}
\alias{polar_save_rdata}
\title{Save all given data as .RData in the \emph{private} \code{tables} directory to which was created for the specific subproject.}
\usage{
polar_save_rdata(...)
}
\arguments{
\item{...}{A collection of variables to store.}
}
\value{
Nothing.
}
\description{
d1 <- data.frame(x=1:10, y=rnorm(10))
d2 <- data.frame(x=1:10, y=runif(10))
polar_save_rdata(d1, d2)
}
