% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_dir.R
\name{polar_save_table_as_rdata}
\alias{polar_save_table_as_rdata}
\title{Save a Table/Data.Frame as .tsv in the \emph{private} \code{tables} directory to which was created for the specific subproject.}
\usage{
polar_save_table_as_rdata(table, table_name = NA)
}
\arguments{
\item{table}{A data.frame/data.table object.}

\item{table_name}{If the default NA is not changed then the name is the name of the table variable.}
}
\value{
Nothing.
}
\description{
d <- data.frame(x=1:10, y=rnorm(10))
polar_save_table_as_tsv(table = d)
}
