% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_dir.R
\name{polar_save_versions}
\alias{polar_save_versions}
\title{Write versions to folder performance}
\usage{
polar_save_versions(versions)
}
\arguments{
\item{versions}{A data.table containing the columns Software and Verion}
}
\description{
Write versions to folder performance
}
