% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_colourise_text.R
\name{rcmd_running}
\alias{rcmd_running}
\title{Check if R CMD is Running}
\usage{
rcmd_running()
}
\description{
This function checks if R CMD is currently running by examining the 'R_TESTS' environment variable.
}
