% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{readExcelFileAsTableList}
\alias{readExcelFileAsTableList}
\title{Read an Excel file and import each sheet as a separate data.table.}
\usage{
readExcelFileAsTableList(excelFile, maxSheetIndex = 1000)
}
\arguments{
\item{excelFile}{The file path to the Excel file.}

\item{maxSheetIndex}{Maximum number of sheets to be read (default is 1000).}
}
\value{
A list of imported sheets as data.tables.
}
\description{
This function reads an Excel file specified by the 'excelFile' parameter
and imports each sheet as a separate data.table. It returns a list of
imported sheets.
}
