% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_helper_functions.R
\name{refreshFhirToken}
\alias{refreshFhirToken}
\title{Refresh FHIR Token}
\usage{
refreshFhirToken()
}
\description{
This function refreshes the FHIR token if it is defined.
}
\details{
If the FHIR_TOKEN is defined, the function attempts to refresh it using the \code{polar_refresh_token} function.
}
