% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{removeRowsWithNAorEmpty}
\alias{removeRowsWithNAorEmpty}
\title{Remove rows with NA or empty values in all (!) specified columns. Emtpy means also string
only consisting of whitespaces.}
\usage{
removeRowsWithNAorEmpty(dt, columns_to_check)
}
\arguments{
\item{dt}{A data.table to process.}

\item{columns_to_check}{A character vector of column names to check for NA or empty values.}
}
\value{
A modified data.table with rows removed based on the specified condition.
}
\description{
This function takes a data.table and a vector of column names as input and returns
a data.table where rows containing NA or empty values in the specified columns are removed.
}
\examples{
# Example data.table
library(data.table)
my_data <- data.table(
  Column1 = c("Hello", "   World   ",   NA, "Example"),
  Column2 = c(      1,            NA,  " ",        NA),
  Column3 = c(    "A",           "B",   "",       "D")
)

# Names of columns to check
columns_to_check <- c("Column1", "Column2", "Column3")

# Remove rows with NA or empty values in specified columns
result_data <- removeRowsWithNAorEmpty(my_data, columns_to_check)
print(my_data)
print(result_data)

}
