% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{replacePatternsInColumn}
\alias{replacePatternsInColumn}
\title{Replace Multiple Patterns in a Specific Column of a Data.Table}
\usage{
replacePatternsInColumn(dt, column_name, patterns_to_replace, replacement)
}
\arguments{
\item{dt}{A data.table containing the target column.}

\item{column_name}{The name of the column to perform replacements in.}

\item{patterns_to_replace}{A vector of patterns to search for in the column's values.}

\item{replacement}{The single replacement string to replace all matched patterns.}
}
\value{
The updated data.table with specified patterns replaced by the single replacement
string in the specified column.
}
\description{
This function replaces multiple patterns with a single replacement string
in a specified column of a data.table.
}
\examples{
library(data.table)

dt <- data.table(
  TextColumn = c("Hello,123world!", "This.is+an_example$123."),
  AnotherExample = c("This is a test.", "Test 123.")
)

# Print the original data.table
print(dt)

# Define patterns to replace as a vector
patterns_to_replace <- c("[0-9]", "[.,+]")

# Define a single replacement string
replacement <- "*"

# Call the function to replace patterns in the "TextColumn" column
dt <- replacePatternsInColumn(dt, "TextColumn", patterns_to_replace, replacement)

# Print the updated data.table
print(dt)

}
