% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_helper_functions.R
\name{run}
\alias{run}
\title{Execute a script with specified message, process, and verbosity level}
\usage{
run(message, process, verbose_level)
}
\arguments{
\item{message}{A character string describing the purpose of the script.}

\item{process}{A function representing the script to be executed.}

\item{verbose_level}{An integer specifying the verbosity level.}
}
\description{
This function runs a script with the provided message and process, controlling
the verbosity level.
}
