% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_table.R
\name{splitColumnToRows}
\alias{splitColumnToRows}
\title{Split a column into separate rows.}
\usage{
splitColumnToRows(dt, columnName, split = "\\\\s+")
}
\arguments{
\item{dt}{The input data.table.}

\item{columnName}{The name of the column to split.}

\item{split}{The regular expression used to split values (default is '\\s+' for whitespace).}
}
\value{
A new data.table with the specified column split into separate rows, while retaining all other columns.
}
\description{
This function takes a data.table and a column name containing strings with whitespace-separated values
and splits each value into a separate row, while retaining all other columns.
}
\examples{
library(data.table)
table2 <- data.table(
  ATC = c('A',     'B',       'C',         'D',     'E'),
  ICD = c('I', 'I0 H0', 'I01 H01', 'I01. H01.', 'I01.2'),
  SOMETHING_ELSE = 1:5
)
print(table2)
splitColumnToRows(table2, "ICD")

}
