% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_str.R
\name{styled_string}
\alias{styled_string}
\title{Styled String}
\usage{
styled_string(
  ...,
  sep = "",
  fg = NULL,
  bg = NULL,
  bold = FALSE,
  italic = FALSE,
  underline = FALSE,
  slowblink = FALSE,
  rapidblink = FALSE,
  invert = FALSE,
  strike_out = FALSE
)
}
\arguments{
\item{...}{Character vectors to be styled.}

\item{sep}{Separator between the styled strings.}

\item{fg}{Foreground color (0-15).}

\item{bg}{Background color (0-15).}

\item{bold}{Use bold text.}

\item{italic}{Use italic text.}

\item{underline}{Underline text.}

\item{slowblink}{Use slow blink effect.}

\item{rapidblink}{Use rapid blink effect.}

\item{invert}{Invert text colors.}

\item{strike_out}{Strike out text.}
}
\value{
A styled string.
}
\description{
Format strings with ANSI escape codes for text styling.
}
