% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_misc.R
\name{verbose_numbers}
\alias{verbose_numbers}
\title{Convert Numbers to Verbose Number Representations}
\usage{
verbose_numbers(n)
}
\arguments{
\item{n}{Numeric vector to be converted.}
}
\value{
A character vector representing the verbose number representation.
}
\description{
This function converts numbers to verbose number representations, such as "1st," "2nd," "3rd," or "th."
}
