% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_helper_functions.R
\name{filterResources}
\alias{filterResources}
\title{Filters the given resources table by the given filter patterns.}
\usage{
filterResources(resources, filter_patterns)
}
\arguments{
\item{resources}{description}

\item{filter_patterns}{description}
}
\description{
Filters the given resources table by the given filter patterns.
}
